refactor: folders for components
