+++ apps/guardian-ui/src/components/dashboard/admin/BalanceCard.tsx
@@ -10,8 +10,8 @@ import {
 } from '@chakra-ui/react';
 import { AuditSummary } from '@fedimint/types';
 import { useTranslation } from '@fedimint/utils';
-import { useAdminContext } from '../hooks';
-import { BalanceTable } from './BalanceTable';
+import { useAdminContext } from '../../../hooks';
+import { BalanceTable } from '../BalanceTable';
 
 export const BalanceCard: React.FC = () => {
   const { t } = useTranslation();
