+++ apps/guardian-ui/src/components/dashboard/gateways/GatewaysCard.tsx
@@ -16,8 +16,8 @@ import {
 import { ClientConfig, Gateway, ModuleKind } from '@fedimint/types';
 import { Table, TableColumn, TableRow } from '@fedimint/ui';
 import { useTranslation, formatEllipsized } from '@fedimint/utils';
-import { ModuleRpc } from '../types';
-import { useAdminContext } from '../hooks';
+import { ModuleRpc } from '../../../types';
+import { useAdminContext } from '../../../hooks';
 import { ReactComponent as InfoIcon } from '../assets/svgs/info.svg';
 
 type TableKey = 'nodeId' | 'gatewayId' | 'fee';
