+++ apps/guardian-ui/src/components/setup/SetupComplete.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Flex, Heading, Text, Button, Icon } from '@chakra-ui/react';
 import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
-import { useAppContext } from '../hooks';
+import { useAppContext } from '../../hooks';
 import { useTranslation } from '@fedimint/utils';
-import { APP_ACTION_TYPE, Status } from '../types';
+import { APP_ACTION_TYPE, Status } from '../../types';
 
 export const SetupComplete: React.FC = () => {
   const { t } = useTranslation();
