+++ apps/guardian-ui/src/components/setup/VerifyGuardians.tsx
@@ -24,11 +24,11 @@ import {
 import { ServerStatus, Peer } from '@fedimint/types';
 import { useTranslation } from '@fedimint/utils';
 import { CopyInput, Table } from '@fedimint/ui';
-import { useConsensusPolling, useSetupContext } from '../hooks';
-import { GuardianRole } from '../types';
+import { useConsensusPolling, useSetupContext } from '../../hooks';
+import { GuardianRole } from '../../types';
 import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
 import { ReactComponent as CopyIcon } from '../assets/svgs/copy.svg';
-import { formatApiErrorMessage } from '../utils/api';
+import { formatApiErrorMessage } from '../../utils/api';
 import { ReactComponent as CheckCircleIcon } from '../assets/svgs/check-circle.svg';
 import { ReactComponent as XCircleIcon } from '../assets/svgs/x-circle.svg';
